+++ js/user-pictures.js
@@ -1,0 +1,18 @@
+import {getPhotos} from './data.js';
+
+const pictureContainer = document.querySelector('.pictures');
+const pictureTemplate = document.querySelector('#picture').content.querySelector('.picture');
+
+const similarPictures = getPhotos();
+
+const similarListFragment = document.createDocumentFragment();
+
+similarPictures.forEach(({url, comments, likes}) => {
+  const pictureElement = pictureTemplate.cloneNode(true);
+  pictureElement.querySelector('.picture__img').src = url;
+  pictureElement.querySelector('.picture__comments').textContent = comments.length;
+  pictureElement.querySelector('.picture__likes').textContent = likes;
+  similarListFragment.appendChild(pictureElement);
+});
+
+pictureContainer.appendChild(similarListFragment);
